+++ src/Atoms/Flag/Flag.tsx
@@ -38,8 +38,14 @@ const FLAGS_WITH_WHITE_BACKGROUND = ['fi', 'jp', 'ru'];
 export const Flag: FlagComponent = ({ country, height, inline }) => {
   assert(Boolean(country), 'Flag: You need to supply a country code');
   const FlagPathComponent = country ? flags[country.toLowerCase()] : null;
+
+  if (!FlagPathComponent) {
+    return null;
+  }
+
   const grayBorder = FLAGS_WITH_WHITE_BACKGROUND.some(R.equals(country.toLowerCase()));
-  return FlagPathComponent ? (
+
+  return (
     <StyledSvg
       inline={inline}
       focusable="false"
@@ -50,5 +56,5 @@ export const Flag: FlagComponent = ({ country, height, inline }) => {
     >
       <FlagPathComponent />
     </StyledSvg>
-  ) : null;
+  );
 };