+++ src/theme/theme.types.ts
@@ -83,8 +83,12 @@ export type Theme = {
     divider: RawColor['gray6'];
     /** gray0 */
     backgroundDark: RawColor['gray0'];
-    /** gray3 */
+    /** @deprecated gray3 */
     disabled: RawColor['gray3'];
+    /** gray3 */
+    disabledText: RawColor['gray3'];
+    /** gray6 */
+    disabledBackground: RawColor['gray6'];
     /** white */
     buttonSecondaryBackground: RawColor['white'];
     /** white */